+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This log is intended to keep track of package changes, including
 but not limited to API changes and file location changes. Minor behavioral
 changes may not be included if they are not expected to break existing code.
 
+## 0.3.7 (2023-02-04)
+
+- Revert parentId from React
+- Fix bug with `inline` mode in React
+
 ## 0.3.6 (2023-02-03)
 
 - Add support for parentId in React SDK.